requirements.txt
google-cloud-bigquery
tink[gcpkms]